Feature/initialize rest api

+++ src/main/java/com/christmas/api/domain/service/ProductTestService.java
@@ -4,6 +4,7 @@ import com.christmas.api.application.IProductTestUseCase;
 import com.christmas.api.domain.ProductTest;
 import com.christmas.api.domain.service.mapper.IProductTestMapper;
 import com.christmas.api.infrastructure.controller.ProductTestDto;
+import com.christmas.api.infrastructure.controller.exceptions.ProductTestNotFoundException;
 import com.christmas.api.infrastructure.repository.ProductTestEntity;
 import com.christmas.api.infrastructure.repository.ProductTestRepository;
 import lombok.RequiredArgsConstructor;
@@ -31,8 +32,9 @@ public class ProductTestService implements IProductTestUseCase {
     @Override
     public ProductTest getProductTestById(Integer id) {
 
-        return iProductTestMapper.toDomain(productTestRepository.getProductTestById(id));
-
+        return productTestRepository.getProductTestById(id)
+                .map(iProductTestMapper::toDomain)
+                .orElseThrow( () -> new ProductTestNotFoundException(id));
     }
 
     @Override
@@ -49,6 +51,9 @@ public class ProductTestService implements IProductTestUseCase {
     @Override
     public ProductTest updateProductTest(Integer id, ProductTestDto productTestDtoToUpdate) {
 
+        if (productTestRepository.getProductTestById(id).isEmpty())
+            throw new  ProductTestNotFoundException(id);
+
         ProductTest domainFromDto = iProductTestMapper.toDomain(productTestDtoToUpdate);
 
         ProductTestEntity entity = productTestRepository.updateProductTest(id, domainFromDto);
@@ -60,6 +65,9 @@ public class ProductTestService implements IProductTestUseCase {
     @Override
     public ProductTest updatePartiallyProductTest(Integer id, ProductTestDto productTestDtoToUpdate) {
 
+        if (productTestRepository.getProductTestById(id).isEmpty())
+            throw new  ProductTestNotFoundException(id);
+
         ProductTest domainFromDto = iProductTestMapper.toDomain(productTestDtoToUpdate);
 
         ProductTestEntity entity = productTestRepository.updateProductTest(id, domainFromDto);
@@ -70,6 +78,9 @@ public class ProductTestService implements IProductTestUseCase {
 
     @Override
     public void deleteProductTestById(Integer id) {
+
+        if (productTestRepository.getProductTestById(id).isEmpty())
+            throw new  ProductTestNotFoundException(id);
 
         productTestRepository.deleteProductTestById(id);
